Added Solver Related Tests to test_solvers.py
My first try at contributing to SymPy.
